+++ Demos/Python/3D/ReconASTRA3D_artifacts.py
@@ -5,10 +5,8 @@ GPLv3 license (ASTRA toolbox)
 Note that the TomoPhantom package is released under Apache License, Version 2.0
 
 * Script to generate 3D analytical phantoms and their projection data using TomoPhantom
-* Synthetic flat fields are also genererated and noise incorporated into data 
-together with normalisation errors. This simulates more challeneging data for 
-reconstruction.
-* tomobar is required for reconstruction
+* Projection data is also generated numerically and reconstructed using 
+* tomobar/ ASTRA TOOLBOX 
 
 >>>>> Dependencies (reconstruction): <<<<<
 1. ASTRA toolbox: conda install -c astra-toolbox astra-toolbox
@@ -24,13 +22,11 @@ import numpy as np
 import tomophantom
 from tomophantom import TomoP3D
 from tomophantom.supp.qualitymetrics import QualityTools
-from tomophantom.supp.flatsgen import flats
-from tomophantom.supp.normraw import normaliser_sim
 
 print ("Building 3D phantom using TomoPhantom software")
 tic=timeit.default_timer()
-model = 9 # select a model number from the library
-N_size = 256 # Define phantom dimensions using a scalar value (cubic phantom)
+model = 13 # select a model number from the library
+N_size = 128 # Define phantom dimensions using a scalar value (cubic phantom)
 path = os.path.dirname(tomophantom.__file__)
 path_library3D = os.path.join(path, "Phantom3DLibrary.dat")
 #This will generate a N_size x N_size x N_size phantom (3D)
@@ -56,7 +52,7 @@ plt.title('3D Phantom, sagittal view')
 plt.show()
 
 # Projection geometry related parameters:
-Horiz_det = int(np.sqrt(2)*N_size) # detector column count (horizontal)
+Horiz_det = int(2*N_size) # detector column count (horizontal)
 Vert_det = N_size # detector row count (vertical) (no reason for it to be > N)
 angles_num = int(0.5*np.pi*N_size); # angles number
 angles = np.linspace(0.0,179.9,angles_num,dtype='float32') # in degrees
@@ -66,7 +62,7 @@ print ("Building 3D analytical projection data with TomoPhantom")
 projData3D_analyt= TomoP3D.ModelSino(model, N_size, Horiz_det, Vert_det, angles, path_library3D)
 
 intens_max = 70
-sliceSel = 150
+sliceSel = int(0.5*N_size)
 plt.figure() 
 plt.subplot(131)
 plt.imshow(projData3D_analyt[:,sliceSel,:],vmin=0, vmax=intens_max)
@@ -79,33 +75,37 @@ plt.imshow(projData3D_analyt[:,:,sliceSel],vmin=0, vmax=intens_max)
 plt.title('Tangentogram view')
 plt.show()
 #%%
-print ("Simulate flat fields, add noise and normalise projections...")
-flatsnum = 20 # generate 20 flat fields
-flatsSIM = flats(Vert_det, Horiz_det, maxheight = 0.01, maxthickness = 1, sigma_noise = 0.2, sigmasmooth = 3, flatsnum=flatsnum)
+print ("Adding noise to projection data")
+from tomophantom.supp.artifacts import _Artifacts_
 
-plt.figure() 
-plt.imshow(flatsSIM[0,:,:],vmin=0, vmax=1)
-plt.title('A selected simulated flat-field')
-#%%
-# Apply normalisation of data and add noise
-flux_intensity = 10000 # controls the level of noise (Poisson) 
-sigma_flats = 500 # control the level of noise in flats (lower creates more ring artifacts)
-projData3D_norm = normaliser_sim(projData3D_analyt, flatsSIM, sigma_flats, flux_intensity)
+# forming dictionaries with artifact types
+_noise_ =  {'type' : 'Poisson',
+            'sigma' : 10000, # noise amplitude
+            'seed' : 0}
+
+_stripes_ = {'percentage' : 1.5,
+             'maxthickness' : 2.0,
+             'intensity' : 0.02,
+             'type' : 'full',             
+             'variability' : 0.01}
+
+projData3D_analyt_noisy = _Artifacts_(projData3D_analyt, _noise_, {}, _stripes_, {})
 
 intens_max = 70
-sliceSel = 150
+sliceSel = int(0.5*N_size)
 plt.figure() 
 plt.subplot(131)
-plt.imshow(projData3D_norm[:,sliceSel,:],vmin=0, vmax=intens_max)
-plt.title('2D Projection (erroneous)')
+plt.imshow(projData3D_analyt_noisy[:,sliceSel,:],vmin=0, vmax=intens_max)
+plt.title('2D noisy Projection (analytical)')
 plt.subplot(132)
-plt.imshow(projData3D_norm[sliceSel,:,:],vmin=0, vmax=intens_max)
-plt.title('Sinogram view')
+plt.imshow(projData3D_analyt_noisy[sliceSel,:,:],vmin=0, vmax=intens_max)
+plt.title('Noisy sinogram view')
 plt.subplot(133)
-plt.imshow(projData3D_norm[:,:,sliceSel],vmin=0, vmax=intens_max)
-plt.title('Tangentogram view')
+plt.imshow(projData3D_analyt_noisy[:,:,sliceSel],vmin=0, vmax=intens_max)
+plt.title('Noisy tangentogram view')
 plt.show()
 #%%
+print ("Reconstruction using FBP from tomobar")
 # initialise tomobar DIRECT reconstruction class ONCE
 from tomobar.methodsDIR import RecToolsDIR
 RectoolsDIR = RecToolsDIR(DetectorsDimH = Horiz_det,  # DetectorsDimH # detector dimension (horizontal)
@@ -113,10 +113,9 @@ RectoolsDIR = RecToolsDIR(DetectorsDimH = Horiz_det,  # DetectorsDimH # detector
                     CenterRotOffset = None, # Center of Rotation (CoR) scalar (for 3D case only)
                     AnglesVec = angles_rad, # array of angles in radians
                     ObjSize = N_size, # a scalar to define reconstructed object dimensions
-                    device = 'gpu')
-#%%
-print ("Reconstruction using FBP from tomobar")
-recNumerical= RectoolsDIR.FBP(projData3D_norm) # FBP reconstruction
+                    device_projector = 'gpu')
+
+recNumerical= RectoolsDIR.FBP(projData3D_analyt_noisy) # FBP reconstruction
 
 sliceSel = int(0.5*N_size)
 max_val = 1
@@ -145,67 +144,44 @@ print ("Reconstructing with FISTA-OS method using tomobar")
 print ("%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%")
 # initialise tomobar ITERATIVE reconstruction class ONCE
 from tomobar.methodsIR import RecToolsIR
-RectoolsIR = RecToolsIR(DetectorsDimH = Horiz_det,  # DetectorsDimH # detector dimension (horizontal)
+Rectools = RecToolsIR(DetectorsDimH = Horiz_det,  # DetectorsDimH # detector dimension (horizontal)
                     DetectorsDimV = Vert_det,  # DetectorsDimV # detector dimension (vertical) for 3D case only
                     CenterRotOffset = 0.0, # Center of Rotation (CoR) scalar (for 3D case only)
                     AnglesVec = angles_rad, # array of angles in radians
                     ObjSize = N_size, # a scalar to define reconstructed object dimensions
                     datafidelity='LS',# data fidelity, choose LS, PWLS (wip), GH (wip), Student (wip)
-                    nonnegativity='ENABLE', # enable nonnegativity constraint (set to 'ENABLE')
-                    OS_number = 12, # the number of subsets, NONE/(or > 1) ~ classical / ordered subsets
-                    tolerance = 1e-07, # tolerance to stop outer iterations earlier
-                    device='gpu')
-lc = RectoolsIR.powermethod() # calculate Lipschitz constant
-#%%
-# Run FISTA reconstrucion algorithm without regularisation
-#RecFISTA = RectoolsIR.FISTA(projData3D_norm, iterationsFISTA = 5, lipschitz_const = lc)
+                    device_projector='gpu')
 
-# Run FISTA reconstrucion algorithm with 3D regularisation
-RecFISTA_reg = RectoolsIR.FISTA(projData3D_norm, 
-                                iterationsFISTA = 8, 
-                                regularisation = 'FGP_TV', 
-                                regularisation_parameter = 0.0007, 
-                                regularisation_iterations = 200,
-                                lipschitz_const = lc)
+# prepare dictionaries with parameters:
+_data_ = {'projection_norm_data' : projData3D_analyt_noisy,
+          'OS_number' : 10} # data dictionary
+lc = Rectools.powermethod(_data_) # calculate Lipschitz constant (run once to initialise)
 
-sliceSel = int(0.5*N_size)
-max_val = 1
-plt.figure() 
-plt.subplot(131)
-plt.imshow(RecFISTA_reg[sliceSel,:,:],vmin=0, vmax=max_val)
-plt.title('3D FISTA-TV Reconstruction, axial view')
+# Run FISTA-OS reconstrucion algorithm without regularisation
+_algorithm_ = {'iterations' : 18,
+               'lipschitz_const' : lc}
 
-plt.subplot(132)
-plt.imshow(RecFISTA_reg[:,sliceSel,:],vmin=0, vmax=max_val)
-plt.title('3D FISTA-TV Reconstruction, coronal view')
+# adding regularisation using the CCPi regularisation toolkit
+_regularisation_ = {'method' : 'PD_TV',
+                    'regul_param' : 0.0005,
+                    'iterations' : 80,
+                    'device_regulariser': 'gpu'}
 
-plt.subplot(133)
-plt.imshow(RecFISTA_reg[:,:,sliceSel],vmin=0, vmax=max_val)
-plt.title('3D FISTA-TV Reconstruction, sagittal view')
-plt.show()
-#%%
-# Run FISTA reconstrucion algorithm with 3D regularisation
-RecFISTA_Huber_TV = RectoolsIR.FISTA(projData3D_norm, 
-                                iterationsFISTA = 8, 
-                                huber_data_threshold = 2.0,
-                                regularisation = 'FGP_TV', 
-                                regularisation_parameter = 0.0007, 
-                                regularisation_iterations = 200,
-                                lipschitz_const = lc)
+RecFISTA_os_reg = Rectools.FISTA(_data_, _algorithm_, _regularisation_)
 
 sliceSel = int(0.5*N_size)
 max_val = 1
 plt.figure() 
 plt.subplot(131)
-plt.imshow(RecFISTA_Huber_TV[sliceSel,:,:],vmin=0, vmax=max_val)
+plt.imshow(RecFISTA_os_reg[sliceSel,:,:],vmin=0, vmax=max_val)
 plt.title('3D FISTA Reconstruction, axial view')
 
 plt.subplot(132)
-plt.imshow(RecFISTA_Huber_TV[:,sliceSel,:],vmin=0, vmax=max_val)
+plt.imshow(RecFISTA_os_reg[:,sliceSel,:],vmin=0, vmax=max_val)
 plt.title('3D FISTA Reconstruction, coronal view')
 
 plt.subplot(133)
-plt.imshow(RecFISTA_Huber_TV[:,:,sliceSel],vmin=0, vmax=max_val)
+plt.imshow(RecFISTA_os_reg[:,:,sliceSel],vmin=0, vmax=max_val)
 plt.title('3D FISTA Reconstruction, sagittal view')
 plt.show()
 #%%